updating indexheap to new accessor functions

+++ math_belief_prop/belief_propagation.h
@@ -373,8 +373,8 @@ public:
   inline int64_t getValL(int id, int x, int y=1, int z=1)            {return *(int64_t*) m_buf[id].getPtr (x, y, z);}  
   inline float   getValF(int id, int x, int y=1, int z=1)            {return *(float*) m_buf[id].getPtr (x, y, z);}  
     
-  inline void   SetValI(int id, float val, int x, int y=1, int z=1)     {*(int32_t*) m_buf[id].getPtr(x, y, z) = val;}
-  inline void   SetValL(int id, float val, int x, int y=1, int z=1)     {*(int64_t*) m_buf[id].getPtr(x, y, z) = val;}  
+  inline void   SetValI(int id, int32_t val, int x, int y=1, int z=1)     {*(int32_t*) m_buf[id].getPtr(x, y, z) = val;}
+  inline void   SetValL(int id, int64_t val, int x, int y=1, int z=1)     {*(int64_t*) m_buf[id].getPtr(x, y, z) = val;}  
   inline void   SetValF(int id, float val, int x, int y=1, int z=1)     {*(float*)   m_buf[id].getPtr(x, y, z) = val;}
